Crate openmls_crypto

OpenMLS Crypto traits

This crate defines traits abstracting the crypto primitives used by OpenMLS. The traits are defined to be used with a key_store implementation but do not necessarily have to use one.

Modules

aead

hash

hkdf

hpke

key_generation

signature

Traits

Supports
Check whether the key store supports certain functionality.