Trait openmls_crypto::aead::Open

pub trait Open {
    type KeyStoreType;
    type KeyStoreIndex;
    type Error;
    type CiphertextTag;
    type Plaintext;
    fn open(
        key_store: &Self::KeyStoreType, 
        key_id: &Self::KeyStoreIndex, 
        aead: AeadType, 
        cipher_text_tag: &Self::CiphertextTag, 
        aad: &[u8], 
        nonce: &[u8]
    ) -> Result<Self::Plaintext, Self::Error>;
}

AEAD Open

Associated Types

type KeyStoreType

The key store type used for Open.

type KeyStoreIndex

The type of the key store id used, i.e. the type for indexing the database.

type Error

The error type returned by Open.

type CiphertextTag

The ciphertext and tag input type of Open::open().

type Plaintext

The plaintext return type of Open::open().

Loading content...
Required methods

fn open(
key_store: &Self::KeyStoreType,
key_id: &Self::KeyStoreIndex,
aead: AeadType,
cipher_text_tag: &Self::CiphertextTag,
aad: &[u8],
nonce: &[u8]
) -> Result<Self::Plaintext, Self::Error>

Decrypt the cipher_text with the given parameters and return the plain text as byte vector.

Loading content...