Trait openmls_crypto::hkdf::HkdfDerive

pub trait HkdfDerive {
    type KeyStoreType;
    type KeyStoreIndex;
    type Secret;
    type Error;
    fn hkdf(
        key_store: &Self::KeyStoreType, 
        ikm: &Self::KeyStoreIndex, 
        hash: HashType, 
        salt: &[u8], 
        info: &[u8], 
        out_len: usize, 
        okm: &Self::KeyStoreIndex
    ) -> Result<(), Self::Error>;

    fn extract(
        key_store: &Self::KeyStoreType, 
        ikm: &Self::KeyStoreIndex, 
        hash: HashType, 
        salt: &[u8], 
        okm: &Self::KeyStoreIndex
    ) -> Result<(), Self::Error>;

    fn expand(
        key_store: &Self::KeyStoreType, 
        prk: &Self::KeyStoreIndex, 
        hash: HashType, 
        info: &[u8], 
        out_len: usize, 
        okm: &Self::KeyStoreIndex
    ) -> Result<(), Self::Error>;

    fn hkdf_with_status(
        key_store: &Self::KeyStoreType, 
        ikm: &Self::KeyStoreIndex, 
        hash: HashType, 
        salt: &[u8], 
        info: &[u8], 
        out_len: usize, 
        okm: &Self::KeyStoreIndex, 
        status: Status
    ) -> Result<(), Self::Error>;

    fn extract_with_status(
        key_store: &Self::KeyStoreType, 
        ikm: &Self::KeyStoreIndex, 
        hash: HashType, 
        salt: &[u8], 
        okm: &Self::KeyStoreIndex, 
        status: Status
    ) -> Result<(), Self::Error>;

    fn expand_with_status(
        key_store: &Self::KeyStoreType, 
        prk: &Self::KeyStoreIndex, 
        hash: HashType, 
        info: &[u8], 
        out_len: usize, 
        okm: &Self::KeyStoreIndex, 
        status: Status
    ) -> Result<(), Self::Error>;
}

HKDF

Associated Types

type KeyStoreType

The key store type used for Seal.

type KeyStoreIndex

The type of the key store id used, i.e. the type for indexing the database.

type Secret

The type of the secrets used in HKDF.

type Error

The error type returned by HkdfDerive.

Loading content...
Required methods

fn hkdf(
key_store: &Self::KeyStoreType,
ikm: &Self::KeyStoreIndex,
hash: HashType,
salt: &[u8],
info: &[u8],
out_len: usize,
okm: &Self::KeyStoreIndex
) -> Result<(), Self::Error>

HKDF Compute HKDF on the input and store it with the okm id.

fn extract(
key_store: &Self::KeyStoreType,
ikm: &Self::KeyStoreIndex,
hash: HashType,
salt: &[u8],
okm: &Self::KeyStoreIndex
) -> Result<(), Self::Error>

HKDF extract Extract pre-key material from ikm and store it with the prk id.

fn expand(
key_store: &Self::KeyStoreType,
prk: &Self::KeyStoreIndex,
hash: HashType,
info: &[u8],
out_len: usize,
okm: &Self::KeyStoreIndex
) -> Result<(), Self::Error>

HKDF expand The expanded secret is stored with the okm id.

fn hkdf_with_status(
key_store: &Self::KeyStoreType,
ikm: &Self::KeyStoreIndex,
hash: HashType,
salt: &[u8],
info: &[u8],
out_len: usize,
okm: &Self::KeyStoreIndex,
status: Status
) -> Result<(), Self::Error>

HKDF Compute HKDF on the input and store it with the okm id.

fn extract_with_status(
key_store: &Self::KeyStoreType,
ikm: &Self::KeyStoreIndex,
hash: HashType,
salt: &[u8],
okm: &Self::KeyStoreIndex,
status: Status
) -> Result<(), Self::Error>

HKDF extract Extract pre-key material from ikm and store it with the prk id with the provided status.

fn expand_with_status(
key_store: &Self::KeyStoreType,
prk: &Self::KeyStoreIndex,
hash: HashType,
info: &[u8],
out_len: usize,
okm: &Self::KeyStoreIndex,
status: Status
) -> Result<(), Self::Error>

HKDF expand The expanded secret is stored with the okm id with the provided status.

Loading content...