Trait openmls_crypto::hpke::HpkeOpen

pub trait HpkeOpen {
    type KeyStoreType;
    type KeyStoreIndex;
    type Plaintext;
    type Ciphertext;
    type KemInput;
    type Error;
    fn hpke_open_with_sk(
        key_store: &Self::KeyStoreType, 
        kdf: KdfType, 
        aead: AeadType, 
        key_id: &Self::KeyStoreIndex, 
        cipher_text: &Self::Ciphertext, 
        kem: &Self::KemInput, 
        info: &[u8], 
        aad: &[u8]
    ) -> Result<Self::Plaintext, Self::Error>;
}

Associated Types

type KeyStoreType

The key store type used for HpkeOpen.

type KeyStoreIndex

The type of the key store id used, i.e. the type for indexing the database.

type Plaintext

The type of the plaintext output.

type Ciphertext

The type of the ciphertext input.

type KemInput

The type of the KEM input.

type Error

The error type returned by HpkeOpen.

Loading content...
Required methods

fn hpke_open_with_sk(
key_store: &Self::KeyStoreType,
kdf: KdfType,
aead: AeadType,
key_id: &Self::KeyStoreIndex,
cipher_text: &Self::Ciphertext,
kem: &Self::KemInput,
info: &[u8],
aad: &[u8]
) -> Result<Self::Plaintext, Self::Error>

Open an HPKE cipher_text with the private key of the given key_id.

Loading content...