Trait openmls_crypto::hpke::HpkeSeal

pub trait HpkeSeal {
    type KeyStoreType;
    type KeyStoreIndex;
    type KemOutput;
    type Ciphertext;
    type Plaintext;
    type PublicKey;
    type Error;
    fn hpke_seal(
        key_store: &Self::KeyStoreType, 
        kdf: KdfType, 
        aead: AeadType, 
        key_id: &Self::KeyStoreIndex, 
        info: &[u8], 
        aad: &[u8], 
        payload: &Self::Plaintext
    ) -> Result<(Self::Ciphertext, Self::KemOutput), Self::Error>;

    fn hpke_seal_to_pk(
        kdf: KdfType, 
        aead: AeadType, 
        key: &Self::PublicKey, 
        info: &[u8], 
        aad: &[u8], 
        payload: &Self::Plaintext
    ) -> Result<(Self::Ciphertext, Self::KemOutput), Self::Error>;

    fn hpke_seal_secret(
        key_store: &Self::KeyStoreType, 
        kdf: KdfType, 
        aead: AeadType, 
        key_id: &Self::KeyStoreIndex, 
        info: &[u8], 
        aad: &[u8], 
        secret_id: &Self::KeyStoreIndex
    ) -> Result<(Self::Ciphertext, Self::KemOutput), Self::Error>;

    fn hpke_seal_secret_to_pk(
        key_store: &Self::KeyStoreType, 
        kdf: KdfType, 
        aead: AeadType, 
        key: &Self::PublicKey, 
        info: &[u8], 
        aad: &[u8], 
        secret_id: &Self::KeyStoreIndex
    ) -> Result<(Self::Ciphertext, Self::KemOutput), Self::Error>;
}

HPKE Note that his trait only holds a very limited subset of HPKE. Only single-shot, base-mode HPKE is supported for now.

Associated Types

type KeyStoreType

The key store type used for HpkeSeal.

type KeyStoreIndex

The type of the key store id used, i.e. the type for indexing the database.

type KemOutput

The type of the KEM output.

type Ciphertext

The type of the ciphertext output.

type Plaintext

The type of the plaintext input.

type PublicKey

The type of the public key input.

type Error

The error type returned by HpkeSeal.

Loading content...
Required methods

fn hpke_seal(
key_store: &Self::KeyStoreType,
kdf: KdfType,
aead: AeadType,
key_id: &Self::KeyStoreIndex,
info: &[u8],
aad: &[u8],
payload: &Self::Plaintext
) -> Result<(Self::Ciphertext, Self::KemOutput), Self::Error>

Encrypt the payload to the public key stored for key_id.

fn hpke_seal_to_pk(
kdf: KdfType,
aead: AeadType,
key: &Self::PublicKey,
info: &[u8],
aad: &[u8],
payload: &Self::Plaintext
) -> Result<(Self::Ciphertext, Self::KemOutput), Self::Error>

Encrypt the payload to the public key.

fn hpke_seal_secret(
key_store: &Self::KeyStoreType,
kdf: KdfType,
aead: AeadType,
key_id: &Self::KeyStoreIndex,
info: &[u8],
aad: &[u8],
secret_id: &Self::KeyStoreIndex
) -> Result<(Self::Ciphertext, Self::KemOutput), Self::Error>

Encrypt the secret stored for secret_id to the public key stored for key_id.

fn hpke_seal_secret_to_pk(
key_store: &Self::KeyStoreType,
kdf: KdfType,
aead: AeadType,
key: &Self::PublicKey,
info: &[u8],
aad: &[u8],
secret_id: &Self::KeyStoreIndex
) -> Result<(Self::Ciphertext, Self::KemOutput), Self::Error>

Encrypt the secret stored for secret_id to the public key.

Loading content...