Trait openmls_crypto::key_generation::GenerateKeys

pub trait GenerateKeys: Supports {
    type KeyStoreType;
    type KeyStoreIndex;
    type PublicKey;
    type Error;
    fn new_secret(
        key_store: &Self::KeyStoreType, 
        key_type: SymmetricKeyType, 
        status: Status, 
        k: &Self::KeyStoreIndex, 
        label: &[u8]
    ) -> Result<(), Self::Error>;

    fn new_key_pair(
        key_store: &Self::KeyStoreType, 
        key_type: AsymmetricKeyType, 
        status: Status, 
        label: &[u8]
    ) -> Result<(Self::PublicKey, Self::KeyStoreIndex), Self::Error>;
}

Generate keys.

Associated Types

type KeyStoreType

The key store type used for GenerateKeys.

type KeyStoreIndex

The type of the key store id used, i.e. the type for indexing the database.

type PublicKey

The type of the public key generated and returned.

type Error

The error type returned by GenerateKeys.

Loading content...
Required methods

fn new_secret(
key_store: &Self::KeyStoreType,
key_type: SymmetricKeyType,
status: Status,
k: &Self::KeyStoreIndex,
label: &[u8]
) -> Result<(), Self::Error>

fn new_key_pair(
key_store: &Self::KeyStoreType,
key_type: AsymmetricKeyType,
status: Status,
label: &[u8]
) -> Result<(Self::PublicKey, Self::KeyStoreIndex), Self::Error>

Generate a new key pair and return the GenerateKeys::PublicKey as well as the identifier of the private key in the key store.

Loading content...